% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_ranger.R
\name{.mtry_grid_calc}
\alias{.mtry_grid_calc}
\title{Select plausible values for mtry to consider}
\usage{
.mtry_grid_calc(p)
}
\arguments{
\item{p}{number of exposures to consider.}
}
\value{
a vector of values of mtry to consider
}
\description{
Select plausible values for mtry to consider
}
