% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_ranger.R
\name{tune_ranger}
\alias{tune_ranger}
\title{Select plausible values for mtry to consider}
\usage{
tune_ranger(
  data,
  exposures,
  outcome,
  weights,
  mtry_grid = NULL,
  min.node.size_grid = c(1, 3, 5, 10),
  num_trees = 500,
  splitrule = "gini",
  n_folds = 5,
  full_results = FALSE,
  include_training_data = FALSE
)
}
\arguments{
\item{data}{dataset}

\item{exposures}{a vector containing the names of the exposure variables in the data}

\item{outcome}{the name of the outcome variable in the data}

\item{weights}{the name of the weights variable in the data}

\item{mtry_grid}{vector of values of mtry to consider. default is c(floor(p/3), floor(sqrt(p)), floor(sqrt(p)*2), floor(sqrt(p)*0.5)) where p is the number of exposures.}

\item{min.node.size_grid}{vector of values of min.node.size to consider. default is 1, 3, 5, and 10}

\item{num_trees}{number of trees to build. default is 500.}

\item{splitrule}{name of splitrule to employ. default is 'gini', which is ranger default for classification and probability programs. consider 'variance' for regression problems.}

\item{n_folds}{number of CV folds}

\item{full_results}{return full results from caret::train(). default is FALSE}

\item{include_training_data}{include training data in output if full_results = TRUE}
}
\value{
a list. either bestTune parameters (default) or output from caret::train()
}
\description{
Select plausible values for mtry to consider
}
