% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wranger.R
\name{wranger}
\alias{wranger}
\title{Weighted random forest prediction models for weighted data via ranger}
\usage{
wranger(
  data = NULL,
  col.y = NULL,
  col.x = NULL,
  cluster = NULL,
  strata = NULL,
  weights = NULL,
  design = NULL,
  family = "binomial",
  mtry_grid = NULL,
  min.node.size_grid = c(1, 3, 5, 10),
  num_trees = 500,
  splitrule = NULL,
  importance = "permutation",
  write.forest = TRUE,
  method = "dCV",
  k = 10,
  R = 1,
  B = 200,
  dCV.sw.test = FALSE,
  train.prob = 0.7,
  method.split = c("dCV", "bootstrap", "subbootstrap"),
  print.rw = FALSE
)
}
\arguments{
\item{data}{A data frame with information about the response variable and covariates, as well as sampling weights and strata and cluster indicators. It could be \code{NULL} if the sampling design is indicated in the \code{design} argument.}

\item{col.y}{A numeric value indicating the number of the column in which information on the response variable can be found or a character string indicating the name of that column.}

\item{col.x}{A numeric vector indicating the numbers of the columns in which information on the covariates can be found or a vector of character strings indicating the names of these columns.}

\item{cluster}{A character string indicating the name of the column with cluster identifiers. It could be \code{NULL} if the sampling design is indicated in the \code{design} argument.}

\item{strata}{A character string indicating the name of the column with strata identifiers. It could be \code{NULL} if the sampling design is indicated in the \code{design} argument.}

\item{weights}{A character string indicating the name of the column with sampling weights. It could be \code{NULL} if the sampling design is indicated in the \code{design} argument.}

\item{design}{An object of class \code{survey.design} generated by \code{survey::svydesign()}. It could be \code{NULL} if information about \code{cluster}, \code{strata}, \code{weights} and \code{data} are given.}

\item{family}{A character string indicating the family to fit LASSO models. Choose between \code{gaussian} (to fit linear models) or \code{binomial} (for logistic models).}

\item{mtry_grid}{A vector of values to try for mtry paramter. Default is c(floor(p/3), floor(sqrt(p)), floor(sqrt(p)*2), floor(sqrt(p)*0.5)) where p is the number of exposures (i.e., length(col.x)).}

\item{min.node.size_grid}{A vector of values to try for min.node.size. Default is c(1, 3, 5, 10).}

\item{num_trees}{Number of trees to grow in forest}

\item{splitrule}{Splitrule for forest. Default is default in ranger based on data.}

\item{importance}{method for selecting variable importance. default is "permutation"}

\item{write.forest}{whether to write the forest model in output. default is TRUE}

\item{method}{A character string indicating the method to be applied to define replicate weights. Choose between one of these: \code{JKn}, \code{dCV}, \code{bootstrap}, \code{subbootstrap}, \code{BRR}, \code{split}, \code{extrapolation}.}

\item{k}{A numeric value indicating the number of folds to be defined. Default is \code{k=10}. Only applies for the \code{dCV} method.}

\item{R}{A numeric value indicating the number of times the sample is partitioned. Default is \code{R=1}. Only applies for \code{dCV}, \code{split} or \code{extrapolation} methods.}

\item{B}{A numeric value indicating the number of bootstrap resamples. Default is \code{B=200}. Only applies for \code{bootstrap} and  \code{subbootstrap} methods.}

\item{dCV.sw.test}{A logical value indicating the method for estimating the error for \code{dCV} method. \code{FALSE}, (the default option) estimates the error for each test set and defines the cross-validated error based on the average strategy. Option \code{TRUE} estimates the cross-validated error based on the pooling strategy}

\item{train.prob}{A numeric value between 0 and 1, indicating the proportion of clusters (for the method \code{split}) or strata (for the method \code{extrapolation}) to be set in the training sets. Default is \code{train.prob = 0.7}. Only applies for \code{split} and \code{extrapolation} methods.}

\item{method.split}{A character string indicating the way in which replicate weights should be defined in the \code{split} method. Choose one of the following: \code{dCV}, \code{bootstrap} or \code{subbootstrap}. Only applies for \code{split} method.}

\item{print.rw}{A logical value. If \code{TRUE}, the data set with the replicate weights is saved in the output object. Default \code{print.rw=FALSE}.}
}
\value{
The output object of the function \code{wranger()} is an object of class \code{wranger}. This object is a list containing 3 or 4 elements, depending on the value set to the argument \code{print.rw}. Below we describe the contents of these elements:
\itemize{
\item \code{param}: A list containing information of two elements:
\itemize{
\item \code{grid}: A numeric vector indicating all the values considered for the tuning parameter.
\item \code{min}: A numeric value indicating the value of the tuning parameters that minimizes the average error (i.e., selected optimal tuning parameter).
}
\item \code{error}: A list containing information of two elements:
\itemize{
\item \code{average}: A numeric vector indicating the average error corresponding to each tuning parameter.
\item \code{all}: A numeric matrix indicating the error of each test set for each tuning parameter.
}
\item \code{model}: A model fitted with optimal tuning paramters
\item \code{data.rw}: A data frame containing the original data set and the replicate weights added to define training and test sets. Only included in the output object if \code{print.rw=TRUE}.
\item \code{call}: an object containing the information about the way in which the function has been run.
}
}
\description{
This function allows as to fit random forest prediction (linear or logistic) models to weighted data, considering sampling weights in the estimation process and selects the mtry and min.node.size that minimizes the error based on different replicating weights methods.
}
\examples{
data(simdata_lasso_binomial)
mcv <- wlasso(data = simdata_lasso_binomial,
              col.y = "y", col.x = 1:50,
              family = "binomial",
              cluster = "cluster", strata = "strata", weights = "weights",
              method = "dCV", k=10, R=20)
}
